Separate `LiveApp` and `HermeticApp` in e2e tests for `CallComposite` (#2065)
